+++ tools/gh_release.py
@@ -21,12 +21,9 @@ def build_tarball(release_dir, release_tgz, arch):
     """Build a release tarball with local assets"""
     # Do not include signatures in GitHub release since we aren't
     # making those keys public.
-    # Exclude CPU templates in GitHub release as they are already
-    # available on GitHub without any action (like building a binary).
     exclude_files = {
         "RELEASE_NOTES",
         "SHA256SUMS.sig",
-        *[f.stem for f in Path("tests/data/custom_cpu_templates").glob("*.json")],
     }
     with tarfile.open(release_tgz, "w:gz") as tar:
         files = [x for x in release_dir.rglob("*") if x.is_file()]